StepGauge renders a top of ReqCard, but each bar displays the same progess not the individual requirement
