+++ src/components/requirement/ReqCard.js
@@ -1,9 +1,12 @@
 import React from "react"
 import { loadFromLocalStorage } from "../../store"
 
-const ReqCard = ({ requirement, fetchSteps }) => {
+import StepGauge from '../../components/steps/StepGauge'
+
+const ReqCard = ({ requirement, fetchSteps, steps }) => {
   return (
     <div className="requirement-card">
+      <StepGauge steps={steps}/>
       <h2 className="title">{requirement.title}</h2>
       <p className="description">{requirement.tasks_description}</p>
       <div
