+++ src/components/steps/StepList.js
@@ -1,18 +1,15 @@
 import React, { useState, useEffect } from "react"
-import { connect } from "react-redux"
-
-import { toggleStep } from "../../actions"
 
 import Step from "./Step"
 import StepGauge from "./StepGauge"
 
 const StepList = props => {
+  console.log(props)
   const [steps, setSteps] = useState([])
-  // console.log(steps)
 
   useEffect(() => {
-    setSteps(props.stepsByTask)
-  }, [props.stepsByTask])
+    setSteps(props.steps)
+  }, [props.steps])
 
   return (
     <div className="step-list-container">
@@ -20,23 +17,11 @@ const StepList = props => {
       <h1 className="title">Steps to complete</h1>
       <div className="step-list">
         {steps.map(step => (
-          <Step key={step.id} step={step} toggle={props.toggleStep} />
+          <Step key={step.id} step={step} toggle={props.toggle} />
         ))}
       </div>
     </div>
   )
 }
 
-const mapStateToProps = state => {
-  return {
-    inProgress: state.stepReducer.inProgress,
-    serverError: state.stepReducer.serverError,
-    isServerError: state.stepReducer.isServerError,
-    stepsByTask: state.stepReducer.stepsByTask
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  { toggleStep }
-)(StepList)
+export default StepList
